Rename "entry" to "index"
